requirements.txt
discord.py
requests
pillow
imageio
numpy